# Denormalize entity in cryptotransferlists and non_fee_transfer tables