--- a/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/CryptoTransfer.java
+++ b/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/CryptoTransfer.java
@@ -23,7 +23,6 @@ package com.hedera.mirror.importer.domain;
 import javax.persistence.Convert;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -34,7 +33,6 @@ import com.hedera.mirror.importer.converter.EntityIdConverter;
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "crypto_transfers")
 public class CryptoTransfer {
     @Id
     private Long consensusTimestamp;