--- a/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/NonFeeTransfer.java
+++ b/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/NonFeeTransfer.java
@@ -20,12 +20,15 @@ package com.hedera.mirror.importer.domain;
  * ‍
  */
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import javax.persistence.Convert;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import com.hedera.mirror.importer.converter.EntityIdConverter;
 
 @Data
 @Entity
@@ -37,9 +40,8 @@ public class NonFeeTransfer {
     @Id
     private Long consensusTimestamp;
 
-    private Long realmNum;
-
-    private Long entityNum;
-
     private Long amount;
+
+    @Convert(converter = EntityIdConverter.class)
+    private EntityId entityId;
 }