--- a/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/NonFeeTransfer.java
+++ b/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/NonFeeTransfer.java
@@ -23,7 +23,6 @@ package com.hedera.mirror.importer.domain;
 import javax.persistence.Convert;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -34,9 +33,8 @@ import com.hedera.mirror.importer.converter.EntityIdConverter;
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "non_fee_transfers")
 public class NonFeeTransfer {
-    // There is not actually a pk on non_fee_transfers.
+    // There is not actually a pk on non_fee_transfer.
     @Id
     private Long consensusTimestamp;