--- a/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/entity/sql/SqlEntityListener.java
+++ b/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/entity/sql/SqlEntityListener.java
@@ -132,7 +132,7 @@ public class SqlEntityListener implements EntityListener, RecordStreamFileListen
                     "VALUES (?, ?, ?, ?, ?) " +
                     "ON CONFLICT DO NOTHING");
 
-            sqlInsertTransferList = connection.prepareStatement("INSERT INTO cryptotransferlists"
+            sqlInsertTransferList = connection.prepareStatement("INSERT INTO crypto_transfers"
                     + " (consensus_timestamp, amount, entity_id)"
                     + " VALUES (?, ?, ?)");