--- a/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/entity/sql/SqlEntityListener.java
+++ b/hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/entity/sql/SqlEntityListener.java
@@ -132,11 +132,11 @@ public class SqlEntityListener implements EntityListener, RecordStreamFileListen
                     "VALUES (?, ?, ?, ?, ?) " +
                     "ON CONFLICT DO NOTHING");
 
-            sqlInsertTransferList = connection.prepareStatement("INSERT INTO crypto_transfers"
+            sqlInsertTransferList = connection.prepareStatement("INSERT INTO crypto_transfer"
                     + " (consensus_timestamp, amount, entity_id)"
                     + " VALUES (?, ?, ?)");
 
-            sqlInsertNonFeeTransfers = connection.prepareStatement("insert into non_fee_transfers"
+            sqlInsertNonFeeTransfers = connection.prepareStatement("insert into non_fee_transfer"
                     + " (consensus_timestamp, amount, entity_id)"
                     + " values (?, ?, ?)");