--- a/hedera-mirror-importer/src/main/resources/db/migration/V1.25.0__denormalize_entities.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/V1.25.0__denormalize_entities.sql
@@ -104,3 +104,45 @@ alter table if exists t_entities
 
 alter table if exists t_entities
     add primary key (id);
+
+-------------------
+-- cryptotransferlists table
+-------------------
+
+create table if not exists cryptotransferlists (
+  entity_id entity_id not null,
+  consensus_timestamp nanos_timestamp not null,
+  amount hbar_tinybars not null
+);
+
+-- Current assumption in t_cryptotransferlists is that shard = 0. Below migration builds on that assumption.
+insert into cryptotransferlists
+select encodeEntityId(0, realm_num, entity_num), consensus_timestamp, amount
+from t_cryptotransferlists;
+
+drop index if exists idx__t_cryptotransferlists__consensus_and_realm_and_num; -- drop indexes explicitly for history
+drop index if exists idx__t_cryptotransferlists__realm_and_num_and_consensus;
+drop table if exists t_cryptotransferlists;
+
+create index if not exists cryptotransferlists__consensus_timestamp
+    on cryptotransferlists (consensus_timestamp);
+create index cryptotransferlists__entity_id
+    on cryptotransferlists (entity_id)
+    where entity_id != 98; -- id corresponding to treasury address 0.0.98
+
+create table if not exists non_fee_transfers_new (
+    entity_id entity_id not null,
+    consensus_timestamp nanos_timestamp not null,
+    amount hbar_tinybars not null
+);
+
+-- Current assumption in non_fee_transfers is that shard = 0. Below migration builds on that assumption.
+insert into non_fee_transfers_new
+select encodeEntityId(0, realm_num, entity_num), consensus_timestamp, amount
+from non_fee_transfers;
+
+drop table if exists non_fee_transfers;
+alter table non_fee_transfers_new rename to non_fee_transfers;
+
+create index non_fee_transfers__consensus_timestamp
+    on non_fee_transfers (consensus_timestamp);