- Use encoded entity id in t_cryptotransferlists table
- Optimize indexes on t_cryptotransferlists table
- Rename t_cryptotransferlists to crypto_transfers
- Performance test: In a db with ~8M rows in t_ctl, and workload which inserted 6000 crypto transfer rows in a batch (no inserts in other tables), time to insert was as follows:
   - Current: 41.4ms
  - With denormalization and non-composite indexes: 33.3ms
  - With partial index: 30ms
  Net gain : 41.4ms --> 30ms
- Partial index
Use partial index for entity_id column of t_cryptotransferlists. Indexing ignores entity ids for the treasury account (0.0.98) since it is present in every transaction. This will also lead to much smaller index since that inserting records with that entity id leads to a lot of btree node splits and half filled Index pages.
- Migration time of db having ~3m txns is 130s. Extrapolating for mainnet (~150m txns), migration will take ~2 hours

Seeing insertions reach 10k at times now:
```
2020-06-18 02:38:16,302 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 100034 transactions from record file ./data/recordstreams/valid/2020-02-09T20_00_00.004108Z.rcd in 17229ms (5806/s)
2020-06-18 02:38:26,158 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 99995 transactions from record file ./data/recordstreams/valid/2020-02-09T20_01_40.009724Z.rcd in 9829ms (10173/s)
2020-06-18 02:38:35,576 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 99997 transactions from record file ./data/recordstreams/valid/2020-02-09T20_03_20.004973Z.rcd in 9405ms (10632/s)
2020-06-18 02:38:45,194 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 100028 transactions from record file ./data/recordstreams/valid/2020-02-09T20_05_00.011814Z.rcd in 9605ms (10414/s)
2020-06-18 02:38:54,424 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 100004 transactions from record file ./data/recordstreams/valid/2020-02-09T20_06_40.019463Z.rcd in 9217ms (10849/s)
2020-06-18 02:39:07,864 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 99998 transactions from record file ./data/recordstreams/valid/2020-02-09T20_08_20.006814Z.rcd in 13427ms (7447/s)
2020-06-18 02:39:24,207 INFO  [scheduling-4] c.h.m.i.p.r.RecordFileParser Finished parsing 100008 transactions from record file ./data/recordstreams/valid/2020-02-09T20_10_00.001180Z.rcd in 16329ms (6124/s)
```
 REST changes:
- param parsing code was too convoluted to update for handling encoded ids. Refactored param parsing code. 
- Removed too many special characters in rest integration test code, were hurting readability when I was fixing test failures. Changed to plain csv.

Signed-off-by: Apekshit Sharma <apekshit.sharma@hedera.com>

**Detailed description**:

**Which issue(s) this PR fixes**:
Fixes #820 

**Special notes for your reviewer**:

**Checklist**
- [ ] Documentation added
- [x] Tests updated